Frontend: better error message handling for FINISH_ABORT in scheduler.py
**Context**
`FINISH_ABORT` is only used in scheduler.py, and it has `is_error=True`.
Currently, the server only logs error for it but the actual message is
not being handled properly to the user.

For instance, when the multi-modality request is too long, it should be
propagated back to the user as a 400. Instead, it will be sent back as a
200 with only 1 completion token.

**This PR**
- Add status_code and err_type to FINISH_ABORT
- Handle 400 status in finish_reason in TokenizerManager