Fix deepseek-v3 with torch.compile in PyTorch 2.6.
In PyTorch 2.6, we erroneously changed the behavior of torch.compile to
coerce all inputs to custom operators to be contiguous. This is wrong
when the custom operator is expecting a non-contiguous Tensor.

In PyTorch 2.7 we have since fixed that. This PR is a workaround for
PyTorch 2.6.

The workaround is applying a tag to the custom operator that says "hey,
torch.compile, please don't coerce the input to be contiguous if it
wasn't contiguous!".

Fixes https://github.com/pytorch/pytorch/issues/150714

Test Plan:
- Ran deepseek-v3 with PyTorch 2.6 and an updated sgl-kernel locally and
  verified that this PR fixed the garbage token issue.