[OAI] Support non-normalized logprobs in OpenAI server
**Context**
Previously, the SGLang backend will normalize the logprobs returned to
the user by applying top_p in `top_p_normalize_probs_torch`.
This is not consistent with OpenAI API. In OpenAI official API, the logprobs
returned to the user is independent of top_p.
This PR intends to fix the gap.

**Changes**
- Add `logprobs_normalization` to `server_args.py` to support OpenAI style
logprobs (Change the default behavior to OpenAI style)
- Keep `logprobs_normalization=auto` to be consistent with the current
SGLang style
- Update `Sampler.forward()` to return logprobs based on `normalization_mode`,
extracted from server_args