[Fix] Reduce busy polling when scheduler is idle
In setups which have long inactivity periods it is desirable to reduce
system power consumption when sglang does nothing. This would lead not
only to power savings, but also to more CPU thermal headroom when a
request eventually comes. This is important in cases when multiple GPUs
are connected as each GPU would otherwise pin one thread at 100% CPU
usage.

The simplest solution is to use zmq.Poller on all sockets that may
receive data that needs handling immediately.