Change with_stack option to use env var
Following the suggestion to use env var
SGLANG_PROFILE_WITH_STACK
with the original `with_stack` option to decide
whether setting stack for pytorch profiler
or not.
The expected behaviour:
If both `with_stack` and SGLANG_PROFLIE_WITH_STACK
are not set, the `with_stack` option is kept as None,
otherwise False if any of these is False.