fp8_gemm_kernel with unaligned K and N (qwen2_5_vl)
For unaligned N, the kernel generate the output tensor with aligned
dim which ideally will not give too much perf penalty

For unaligned K, the kernel has to initialize temporary new tensor
for both A and B and performs two additional copies. There are 2
alternatives here:
1. In weight initialization, initialize weight matrix to have a hidden
dim which is 16 aligned with 0 padding for qwen2.5 VL 7B
2. Disable quantization for vision part

Feedback is welcome for which one of three to use.