Fix DeepSeek torch.compile caching link
The reference page for DeepSeek models has a section mentioning using
`torch.compile` and recommending caching the results to speed up
subsequent boot times. The link for caching details pointed at the
general optimizing page that has since been updated to no longer have
any details on caching.

This updates the link to point to the dedicated torch compile caching
reference page, as well as a slight polish of an upstream link included
on that page.

Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>