afd: Add A-F scheduler sync for request process
A new socket is created for A-F control request sync, i.e., control
requests are forwarded from ATTN for FFN so that FFN can also
respond to them. Note that we do not collect the FFN response at
ATTN.

The socket is only active at tp_rank 0, and other ranks get the
requests by broadcasting.

For work requests, everytime ATTN generates a new batch, it sends an
AFSyncReq to FFN to notify that FFN should start model forwarding.
AFSyncReq can be seen as a kind of control request.

Besides, warmup is disabled at FFN so that no work request arrives
at the FFN side.

Signed-off-by: Philo Lu <lulie@linux.alibaba.com>