afd: Model sharding
This patch introduces AFD sharding support for MoE models. It enables
decoupled execution of attention and MoE. AFDProxyAttention is added as
an attention placeholder for the FFN side. For FFN, only experts are
reserved, which will be replaced with AFDMoE in following patches.

In AFD mode, weight loading is optimized by role, i.e., attention
workers load only the attention weights, and so do FFN workers.

Currently the changes only applys to Qwen3MoE.

Signed-off-by: Philo Lu <lulie@linux.alibaba.com>
Co-authored-by: D. Wythe <alibuda@linux.alibaba.com>