afd: introduce AFDMoE
This commit introduces the AFDMoE class, but its implementation is split
into two parts — one for the attention (attn) side and one for the FFN
side.

The reason for this split is that, on the attention side, we do not
further integrate FuseMoE, because no MoE model parameters need to be
loaded there. The workload on the attention side is also very simple: it
only involves sending out data and receiving results back from the FFN
side — no actual computation is required. Therefore, the attention-side
class directly inherits from torch.nn.Module.

In contrast, the FFN side must perform MoE computation, so it inherits
from FuseMoE.

AFDDispatcher serves as an abstract base class for all dispatchers under
AFD mode. Currently, it only supports integration with StepMesh, but in
theory, it can be extended in the future to support additional
communication libraries.

When AFD is enabled, we require the data passed to the MLP to be in
scatter mode.

Signed-off-by: Xuan Zhuo <xuanzhuo@linux.alibaba.com>