Consolidate parser checks for chat completions
**Context**
In processing.rs, we use `factory.get_pooled` to get the corresponding
tool parser and reasoning parser for a model.

However, the fallback behavior for a model that is not found in the parser
registry are different in ToolParser and ReasoningParser. ToolParser
fallbacks to `json`. ReasoningParser fallbacks to a passthrough with no
op.

We should make the behavior consistent. And better yet, we can follow
streaming.rs to check if a model has parsers in processing.rs

**Changes**
- Add a `PassthroughParser` in tool_parsers.
- Use `PassthroughParser` as the default option for tool parsers.
- Add `check_reasoning_parser_availability` and `check_tool_parser_availability`
in utils.rs
- Use `check_reasoning_parser_availability` and `check_tool_parser_availability`
in streaming.rs and processing.rs