[Bugfix] when use cuda graph, the accept rate of MTP2 is about doubled by this fix
The dtype of ForwardBatch.seq_lens_cpu is long.  The data type of kv_len_arr of BatchMLAPagedAttentionWrapper.plan is int32. 
So a long type is used as the arg of BatchMLAPagedAttentionWrapper.plan , which causes half of the attention result is all 0. So the MTP2 accept rate is low when use cuda graph. 

This PR fix this bug.